[ORCHESTRATION] Refactor CLI entry
