+++ tests/test_main_cleanup.py
@@ -3,6 +3,7 @@ from types import SimpleNamespace
 from unittest.mock import AsyncMock
 
 import main
+from orchestration import cli_runner
 
 
 def test_main_invokes_shutdown(monkeypatch):
@@ -12,7 +13,7 @@ def test_main_invokes_shutdown(monkeypatch):
         shutdown=AsyncMock(),
     )
 
-    monkeypatch.setattr(main, "NANA_Orchestrator", lambda: orchestrator)
+    monkeypatch.setattr(cli_runner, "NANA_Orchestrator", lambda: orchestrator)
     monkeypatch.setattr(sys, "argv", ["prog"])
 
     main.main()
